Yara_scanner/requirements.txt
yara-python
colorama
tqdm
reportlab
